requirements.txt
asyncpg==0.31.0
fastapi==0.136.0
google-genai==1.73.1
nodriver==0.48.1
openai==2.32.0
SQLAlchemy==2.0.49
uvicorn[standard]==0.45.0
